Use `max_by_key()` to find best individual
This replaces my somewhat clumsy use of `fold()` with the much tidier `max_by_key()` (which I hadn't realized existed) to find the best individual.

In making this change, though, lots of things broke because it turns out that `f64` implements the `Partial_ord` trait, but not the `Ord` trait, since things like NaN and infinity aren't directly comparable to all values.

In other words, Rust is preventing me from doing something stupid here (which would probably make a nice video or blog post).

I dealt with this by replacing all my `f64` fitnesses with `i64` values instead, which made the world happy.

+++ src/bitstring.rs
@@ -6,8 +6,8 @@ use crate::population::Population;
 
 pub type Bitstring = Vec<bool>;
 
-pub fn count_ones(bits: &Bitstring) -> f64 {
-    bits.iter().filter(|&&bit| bit).count() as f64
+pub fn count_ones(bits: &Bitstring) -> i64 {
+    bits.iter().filter(|&&bit| bit).count() as i64
 }
 
 pub fn make_bitstring(len: usize, rng: &mut ThreadRng) -> Bitstring {
@@ -15,7 +15,7 @@ pub fn make_bitstring(len: usize, rng: &mut ThreadRng) -> Bitstring {
 }
 
 impl Individual<Bitstring> {
-    pub fn new_bitstring(bit_length: usize, compute_fitness: impl Fn(&Bitstring) -> f64, rng: &mut ThreadRng) -> Individual<Bitstring> {
+    pub fn new_bitstring(bit_length: usize, compute_fitness: impl Fn(&Bitstring) -> i64, rng: &mut ThreadRng) -> Individual<Bitstring> {
         Individual::new(
                 |rng| make_bitstring(bit_length, rng), 
                 compute_fitness,
