+++ src/individual.rs
@@ -3,13 +3,13 @@ use rand::rngs::ThreadRng;
 #[derive(Debug)]
 pub struct Individual<T> {
     pub genome: T,
-    pub fitness: f64,
+    pub fitness: i64,
 }
 
 impl<T> Individual<T> {
     pub fn new(
             mut make_genome: impl FnMut(&mut ThreadRng) -> T, 
-            compute_fitness: impl Fn(&T) -> f64,
+            compute_fitness: impl Fn(&T) -> i64,
             rng: &mut ThreadRng) 
         -> Individual<T>
     {
